Removing plugins I haven't used in a while
I am probably going to miss ultisnips, but I think I'd like to try to
actually write a little bit more and not rely on it to remind me about
whether a language has `switch` or `case`.

Also removes errors related to whether my vim has python enabled or not.
:P