Enable italic comments using kitty
After digging around, I figure that this is the easiest way to handle
this. I'm using the extended xterm protocol that kitty provides so that
I don't have to use a makefile or script to install it on every
computer. This breaks it being cross platform to anything that I'm not
using kitty on, but that's ok for now.

References:
https://sw.kovidgoyal.net/kitty/protocol-extensions.html