Feat: Enable direnv
I use LSP programs for nvim... but I don't need all of them running at
the same time. This lets me run a shell with
solargraph/lua-language-server only when I enter a folder that _needs_ it.

(It's still installed somewhere in the nix-store, I'm just making a bet
that I'll speed up nvim or be able to override it's packages too.)